Renaming classes, removing service and engine suffix.
